ptrvec.c: fix empty vector assigning, SYS-1495
(src->size_ + NDM_PTRVEC_BLOCK_SIZE_ - 1) & ~(NDM_PTRVEC_BLOCK_SIZE_ - 1)
gives a zero value when src->size_ is zero but the insert function suppose
that when src->data_ is not NULL it points to an allocated memory block of
NDM_PTRVEC_BLOCK_SIZE_ elements at least.

So the following code

    struct ndm_ptrvec_t a = NDM_PTRVEC_INITIALIZER;
    struct ndm_ptrvec_t b = NDM_PTRVEC_INITIALIZER;

    if (!ndm_ptrvec_assign(&a, &b) || !ndm_ptrvec_insert(&a, 0, NULL)) {
        ...
    }

may lead to UB if an implementation of malloc(0) returns a non-NULL pointer.